Added: Function Pixel.CreateSnake(): creates a snake in the center Function Pixel.GenerateApple(): generates an apple at random location, not where snake is. Thread for while loop in order to be able to press key buttons while graphics are rendering. Function PaintBoard() and drawRectangle: they draw the board variable wait, x_offshoot, y_offshoot with comments Image explanation for CreateSnake() variable start_length in class Snake label showing points earned (not functioning yet) 3rd if statement in Snake.Alive(): holds condition if snake touch itself
Changed:
corrected value of directions in ProcessCmdKey()
moved while.loop from btnPlay_Click() to thread()
method of assigning random place for apple
parameters in Pixel class constructror
if-condition in Pixel.Refresh to 'This element is a snake'
if-conditions now check if input is opposite of current direction, if so then condition is false and nothing happens

Removed:
variable timer in Pixel.GenerateApple() along with for-loops surrounding the assigning of apple.
unnecessary code for painting at bootm of class Snake.

The game is now fundamentally functional! More features need creating still, like points and settings menu in-game.